+++ pkg/providers/dns/testdata/TestResDnsRecord/quotation_marks/create_txt_ml.tf
@@ -6,6 +6,7 @@ resource "akamai_dns_record" "txt_record" {
   zone       = "infrastructure.domain.net"
   name       = "infrastructure.domain.net"
   recordtype = "TXT"
-  target = ["v=spf1 mx include:spf.domain.com include:spf.protection.outlook.com -all"]
-  ttl = 1800
+  target     = ["\"v=spf1 mx include:spf.domain.com include:spf.protection.outlook.com -all\"", "foo foo"]
+  ttl        = 1800
 }
+