+++ pkg/providers/dns/testdata/TestResDnsRecord/quotation_marks/create_txt_sl.tf
@@ -7,6 +7,6 @@ resource "akamai_dns_record" "txt_record" {
   name       = "infrastructure.domain.net"
   recordtype = "TXT"
   target     = ["\"v=spf1 mx include:spf.domain.com include:spf.protection.outlook.com -all\""]
-  ttl = 1800
+  ttl        = 1800
 }